Moved Node to its own file
