+++ lluvia/cpp/core/include/lluvia/core/ComputeNodeDescriptor.h
@@ -8,7 +8,7 @@
 #ifndef LLUVIA_CORE_COMPUTE_NODE_DESCRIPTOR_H_
 #define LLUVIA_CORE_COMPUTE_NODE_DESCRIPTOR_H_
 
-#include "lluvia/core/Node.h"
+#include "lluvia/core/node/Node.h"
 #include "lluvia/core/Parameter.h"
 #include "lluvia/core/PushConstants.h"
 #include "lluvia/core/types.h"
