+++ lluvia/cpp/core/src/node/PortType.cpp
@@ -1,22 +1,24 @@
 /**
-@file       Node.cpp
-@brief      Node class.
+@file       PortType.cpp
+@brief      PortType class.
 @copyright  2018, Juan David Adarve Bermudez. See AUTHORS for more details.
             Distributed under the Apache-2 license, see LICENSE for more details.
 */
 
-#include "lluvia/core/Node.h"
+#include "lluvia/core/node/PortType.h"
 
 #include "lluvia/core/error.h"
 
 #include <exception>
 
-namespace ll {
+namespace ll
+{
 
+    vk::DescriptorType portTypeToVkDescriptorType(const ll::PortType &param)
+    {
 
-vk::DescriptorType portTypeToVkDescriptorType(const ll::PortType& param) {
-
-    switch (param) {
+        switch (param)
+        {
         case ll::PortType::Buffer:
             return vk::DescriptorType::eStorageBuffer;
 
@@ -25,13 +27,14 @@ vk::DescriptorType portTypeToVkDescriptorType(const ll::PortType& param) {
 
         case ll::PortType::SampledImageView:
             return vk::DescriptorType::eCombinedImageSampler;
+        }
     }
-}
 
+    ll::PortType vkDescriptorTypeToPortType(const vk::DescriptorType &vkDescType)
+    {
 
-ll::PortType vkDescriptorTypeToPortType(const vk::DescriptorType& vkDescType) {
-
-    switch (vkDescType) {
+        switch (vkDescType)
+        {
         case vk::DescriptorType::eStorageBuffer:
             return ll::PortType::Buffer;
 
@@ -51,50 +54,7 @@ ll::PortType vkDescriptorTypeToPortType(const vk::DescriptorType& vkDescType) {
         case vk::DescriptorType::eInputAttachment:
         default: // to cover descriptor types added by extensions
             throw std::system_error(createErrorCode(ll::ErrorCode::EnumConversionFailed), "cannot convert from Vulkan DescriptorType enum value to ll::PortType.");
+        }
     }
-}
-
-
-void Node::setState(const ll::NodeState tState) {
-
-    // state transition
-    switch(m_state) {
-        case ll::NodeState::Created:
-        
-            switch(tState) {
-                case ll::NodeState::Created:
-                    // do nothing
-                break;
-                case ll::NodeState::Init:
-                    m_state = tState;
-                    onInit();
-                break;
-            }
-        break;
-
-        case ll::NodeState::Init:
-
-            switch(m_state) {
-                case ll::NodeState::Created:
-                    // TODO: what to do?
-                break;
-                case ll::NodeState::Init:
-                    // TODO: throw exception, node already in init state
-                    throw std::runtime_error("Node already in init state");
-                break;
-            }
-        break;
-    }
-
-}
-
-ll::NodeState Node::getState() const noexcept {
-    return m_state;
-}
-
-
-void Node::init() {
-    setState(ll::NodeState::Init);
-}
 
 } // namespace ll
