+++ lluvia/python/src/lluvia/core/node_new/node_state.pxd
@@ -1,8 +1,8 @@
 """
-    lluvia.core.node
-    ----------------
+    lluvia.core.node.node_state
+    ---------------------------
 
-    :copyright: 2018, Juan David Adarve Bermudez. See AUTHORS for more details.
+    :copyright: 2022, Juan David Adarve Bermudez. See AUTHORS for more details.
     :license: Apache-2 license, see LICENSE for more details.
 """
