+++ lluvia/python/src/lluvia/core/node_new/port_type.pxd
@@ -2,7 +2,7 @@
     lluvia.core.node.port_type
     --------------------------
 
-    :copyright: 2018, Juan David Adarve Bermudez. See AUTHORS for more details.
+    :copyright: 2022, Juan David Adarve Bermudez. See AUTHORS for more details.
     :license: Apache-2 license, see LICENSE for more details.
 """
